requirements.txt
# Framework web
flask>=3.0.0

# OpenAI
openai>=1.12.0

# Firebase
firebase-admin>=6.0.0

# Extraction de texte
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.0.0

# Utilitaires
werkzeug>=3.0.0
python-dotenv>=1.0.0

# Pour Vercel (optionnel, local dev)
gunicorn>=21.2.0
